+++ src/Block/Type/Breadcrumbs.php
@@ -45,6 +45,7 @@ final class Breadcrumbs implements Block
 				'labels'         => $this->attributes['labels']         ?? []
 			],
 			markupConfig: [
+				'namespace'     => 'wp-block-x3p0-breadcrumbs',
 				'containerAttr' => $this->getWrapperAttributes(),
 				'showOnFront'   => $this->attributes['showOnHomepage'] ?? false,
 				'showFirstItem' => $this->attributes['showTrailStart'] ?? true,
@@ -63,8 +64,7 @@ final class Breadcrumbs implements Block
 	 */
 	private function getWrapperAttributes(): array
 	{
-		// Set up some default class names.
-		$classes = ['breadcrumbs' => 'breadcrumbs'];
+		$classes = [];
 
 		// If there is a selected home icon, define the class.
 		if (